remove handler interface
